requirements.txt
crewai==0.28.8
langchain_cohere==0.1.2
langchain_community==0.0.32
langchain_groq==0.1.2
langchain_text_splitters==0.0.1
python-dotenv==1.0.1
streamlit==1.33.0
